Relocate the K-Sized Windows problem under the Deques section
